+++ src/common/memory.h
@@ -67,7 +67,7 @@ class PageAllocator {
     FreeAll();
   }
 
-  void *Alloc(unsigned bytes) {
+  void *Alloc(size_t bytes) {
     if (!bytes)
       return NULL;
 
@@ -82,7 +82,7 @@ class PageAllocator {
       return ret;
     }
 
-    const unsigned pages =
+    const size_t pages =
         (bytes + sizeof(PageHeader) + page_size_ - 1) / page_size_;
     uint8_t *const ret = GetNPages(pages);
     if (!ret)
@@ -109,7 +109,7 @@ class PageAllocator {
   }
 
  private:
-  uint8_t *GetNPages(unsigned num_pages) {
+  uint8_t *GetNPages(size_t num_pages) {
 #ifdef __x86_64
     void *a = sys_mmap(NULL, page_size_ * num_pages, PROT_READ | PROT_WRITE,
                        MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
@@ -139,7 +139,7 @@ class PageAllocator {
 
   struct PageHeader {
     PageHeader *next;  // pointer to the start of the next set of pages.
-    unsigned num_pages;  // the number of pages in this set.
+    size_t num_pages;  // the number of pages in this set.
   };
 
   const unsigned page_size_;
